mitrans-unfixed

+++ SERVER/models/transaction.go
@@ -1,10 +1,10 @@
 package models
 
 type Transaction struct {
-	ID     int64  `json:"id"`
+	ID     int    `json:"id"`
 	UserID int    `json:"user_id"`
 	User   User   `json:"user"`
 	Status string `json:"status"`
-	Total  int    `json:"total"`
+	Total  int64  `json:"total"`
 	Carts  []Cart `json:"carts"`
 }
